minor language revision.
Since the banner won't just be visible on sub menu items but every page that's using this page's itemId.
i.e if this is a catrgory layout. selection the later option would make the banner visible on all articles of the category (unless they are linked to a menu item directly), technically they are not sub menu items but kind of "linked menu items" is the term we are looking for.