-alias appended to the alias class in body
Usually if you have your menu alias as about-us, it will be included as a class in the body but that seems to create problem for some common words/pages such as login or signup.
So, starting this version the class would be {menu-alias}-alias. in the about case about-us-alias, If you have custom CSS using the menu alias classes, please move them to use the -alias appended as the menu alias from body class (without -alias appeneded) will be removed in the next version.